refactor : refactor rename constant package
